Register ContentSettingsTypes as persistent and drop BraveEphemeralProvider
Migrating away from ContentSettingsType::PLUGIN and ResourceIdentifier
means that we no longer need to use ephemeral storage for Brave Shields
supported modes, as we now register our own Brave-specific types for
them and we can make them use persistent, right from the start.

This means that we don't need to have a BraveEphemeralProvider subclass
only to make sure that we don't handle Brave Shields in the same way
the Flash plugin (now removed) used to be handled (i.e. ephemeral), so
we can now remove that class along with the bits in BravePrefProvider's
constructor that would register Brave Shields, since that will now be
handled by upstream's PrefProvider constructor already.