[Android] Updated values for BraveShieldsContentSettings
To get in line with values in 'components/brave_shields/common/brave_shield_constants.h'

Related to our change:
https://github.com/brave/brave-core/commit/434ceca5417b6daf03f920e71bc40c91d5663de0

Remove "bs_" prefix from settings names and modify only those needed

In a previous commit we added a "bs_" (standing for "Brave Shields")
to all the settings' names for consistency, even if only "ads", cookies"
and "javascript" needed to be changed to prevent collissions with other
types from upstream when registering the types.

However this is problematic and confusing, so let's only modify the
ones strictly needed, that is, "cookies" and "ads". For "javascript"
we can simply stop registering our own ContentSettingsType for that
since the one that's actually being used is upstream's.